requirements.txt
manim
openai
python-dotenv
ffmpeg-python 